Split HttpServer finally into many filse
