+++ conf/default.conf
@@ -28,7 +28,7 @@ server {
 
 	root html/server2;
 
-	index index.html;
+	index xx aa index.html bb;
 
 	location / {
 		limit_except POST;
